requirements.txt
absl-py==1.4.0
aio-pika==8.2.3
aiofiles==24.1.0
aiogram==2.25.2
aiohttp==3.8.6
aiohttp-retry==2.9.1
aiormq==6.4.2
aiosignal==1.3.1
APScheduler==3.9.1.post1
astroid==3.2.4
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
Babel==2.9.1
backports.zoneinfo==0.2.1
bidict==0.23.1
blinker==1.8.2
boto3==1.37.38
botocore==1.37.38
CacheControl==0.12.14
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
confluent-kafka==2.11.0
cryptography==45.0.5
cycler==0.12.1
dask==2022.10.2
dill==0.4.0
dnspython==2.3.0
docopt==0.6.2
fbmessenger==6.0.0
filelock==3.16.1
fire==0.7.0
flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.0
future==1.0.0
gast==0.4.0
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.16.0
h5py==3.11.0
httptools==0.6.4
huggingface-hub==0.34.1
humanfriendly==10.0
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
isort==5.13.2
itsdangerous==2.2.0
jax==0.4.13
jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jsonpickle==3.0.4
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.7
libclang==18.1.1
locket==1.0.0
magic-filter==1.0.12
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.5.3
mattermostwrapper==2.2
mccabe==0.7.0
ml-dtypes==0.2.0
msgpack==1.1.1
multidict==5.2.0
mysql-connector-python==9.0.0
networkx==2.6.3
numpy==1.23.5
oauthlib==3.3.1
opt-einsum==3.4.0
packaging==20.9
pamqp==3.2.1
partd==1.4.1
pillow==10.4.0
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
prompt-toolkit==3.0.28
propcache==0.2.0
protobuf==4.23.3
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==1.10.9
pydot==1.4.2
PyJWT==2.9.0
pykwalify==1.8.0
pylint==3.2.7
pymongo==4.3.3
pyparsing==3.1.4
pyreadline3==3.5.4
pyrsistent==0.20.0
python-crfsuite==0.9.11
python-dateutil==2.8.2
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2022.7.1
pywin32==311
PyYAML==6.0.2
questionary==1.10.0
randomname==0.1.5
rasa==3.6.21
rasa-sdk==3.6.2
redis==4.6.0
regex==2022.10.31
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rocketchat-API==1.30.0
rsa==4.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
s3transfer==0.11.5
safetensors==0.4.5
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.10.1
sentry-sdk==1.14.0
simple-websocket==1.1.0
six==1.17.0
sklearn-crfsuite==0.3.6
skops==0.9.0
slack-sdk==3.36.0
SQLAlchemy==1.4.54
structlog==23.3.0
structlog-sentry==2.1.0
tabulate==0.9.0
tarsafe==0.0.4
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tomli==2.2.1
tomlkit==0.13.3
toolz==1.0.0
tqdm==4.67.1
twilio==8.2.2
typing-extensions==4.13.2
typing-utils==0.1.0
tzdata==2025.2
tzlocal==5.2
ujson==5.10.0
urllib3==1.26.20
wcwidth==0.2.13
webexteamssdk==1.6.1
websockets==10.4
werkzeug==3.0.6
wrapt==1.14.1
wsproto==1.2.0
yarl==1.15.2
zipp==3.20.2
